Allow custom HTTP client for Connection
I've added the ability to specify a custom HTTP client when creating a Connection. This is important because the various Azure DevOps clients make requests during construction to get the resource area info. This meant that some requests would be made with the default HTTP client even when specifying a custom one for the git client.